RV64 floating point fused-multiply-add
Replace re-use of the floating-point multiplication and addition units to do fused-multiply-add operations with a dedicated floatfused unit.More accurate as the results of the multiplication are not rounded before use in the addition/subtraction operation.

Saves cycles as the rounding and combining step is eliminated after the multiplication

Chnage wait-state system in floataddsub floatmultiply.

+++ RV64/ROM/BIOS/c/VBIOS.c
@@ -445,10 +445,10 @@ void SMTSTART( unsigned int code ) {
 
 void smtmandel( void ) {
     const int graphwidth = 320, graphheight = 100;
-    double kt = 63, m = 4.0;
-    double xmin = -2.1, xmax = 0.6, ymin = -1.35, ymax = 1.35;
-    double dx = (xmax - xmin) / graphwidth, dy = (ymax - ymin) / graphheight;
-    double jx, jy, tx, ty, wx, wy, r;
+    float kt = 63, m = 4.0;
+    float xmin = -2.1, xmax = 0.6, ymin = -1.35, ymax = 1.35;
+    float dx = (xmax - xmin) / graphwidth, dy = (ymax - ymin) / graphheight;
+    float jx, jy, tx, ty, wx, wy, r;
     int k;
 
     gpu_pixelblock_start( 0, 122, 320 );
